backend/ml_service/requirements.txt
fastapi
uvicorn
google-cloud-storage
scikit-learn
numpy
joblib
xgboost==1.7.6
